+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -15,15 +15,15 @@ public interface UserStorage {
 
     int generatorId();
 
-    public User getUser(int id);
+    public User getUser(long id);
 
-    public Set<Long> addFriend(int idUser, int idFriend);
+    public Set<Long> addFriend(long idUser, long idFriend);
 
-    public Set<Long> deleteFriend(int idUser, int idFriend);
+    public Set<Long> deleteFriend(long idUser, long idFriend);
 
-    public List<User> showFriends(int idUser);
+    public List<User> showFriends(long idUser);
 
-    public List<User> showCommonFriends(int userId, int otherId);
+    public List<User> showCommonFriends(long userId, long otherId);
 
-    public HashMap<Integer, User> getUsers();
+    public HashMap<Long, User> getUsers();
 }
